Refactor HTTP DomainClient for pluggable serialization
Refactored the HTTP-based DomainClient infrastructure to support pluggable serialization formats (binary and XML) via new abstractions. Introduced DataContractHttpDomainClient and HttpDomainClientFactory base classes to centralize request/response and HttpClient logic. Added XmlHttpDomainClient and XmlHttpDomainClientFactory for XML support. Renamed and generalized the serialization helper. Removed duplicated code, making the system more modular and maintainable.
